Allow selection of code from main.py in notebooks
For certain exercises it may be required to have a single main.py which
is split automatically by test_exercises into multiple notebook cells
based on linebreaks. This functionality works fine, but in some
circumstances we may have a single main.py which actually needs split
amongst several notebooks. Now, the main.py can be annotataed with
headers which determine which code cells (i.e. single blocks in main.py)
need to be inserted into which notebooks. So if notebook1.ipynb has two
code cells which correspond to the first two blocks in main.py, and
notebook2.ipynb requires only the second two, main.py would look like

   #NOTEBOOK notebook1.ipynb
   ...

   #NOTEBOOK notebook1.ipynb notebook2.ipynb
   ...

   #NOTEBOOK notebook2.ipynb
   ...

This allows the command line testing of a single main.py file with
`pytest` and the testing of the necessary notebooks with

   pytest --fname notebook1.ipynb <path to>/tests/test_exercises.py