Update marking utilities to use canvas_selector
I've cleaved off some functionality into a separate package for
interacting with the canvas_api from the command line. It's called
canvas_selector.

This update updates grade_ipynbs to use that package.

It also means that the APIKEY.py is no longer required. The information
is kept in a configuration file .canvasapirc instead. For using the
docker image it will need to be copied into the marking directory.

The documentation needs to be updated to reflect this